# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "家庭点菜 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/dishes": {
      "get": {
        "summary": "Get Dishes",
        "operationId": "get_dishes_api_dishes_get",
        "parameters": [
          {
            "name": "cook_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cook Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishRead"
                  },
                  "title": "Response Get Dishes Api Dishes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Dish",
        "operationId": "post_dish_api_dishes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dishes/{dish_id}": {
      "get": {
        "summary": "Get Dish",
        "operationId": "get_dish_api_dishes__dish_id__get",
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dish Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Dish",
        "operationId": "patch_dish_api_dishes__dish_id__patch",
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dish Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Dish",
        "operationId": "delete_dish_api_dishes__dish_id__delete",
        "parameters": [
          {
            "name": "dish_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dish Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingredients": {
      "get": {
        "summary": "Get Ingredients",
        "operationId": "get_ingredients_api_ingredients_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IngredientRead"
                  },
                  "title": "Response Get Ingredients Api Ingredients Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Ingredient",
        "operationId": "post_ingredient_api_ingredients_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/events": {
      "post": {
        "summary": "Create Event",
        "operationId": "create_event_api_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/validation-checkins/{week_start}": {
      "put": {
        "summary": "Put Validation Checkin",
        "operationId": "put_validation_checkin_api_validation_checkins__week_start__put",
        "parameters": [
          {
            "name": "week_start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Week Start"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationCheckinWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationCheckinRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/history": {
      "get": {
        "summary": "Get History",
        "operationId": "get_history_api_history_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  },
                  "title": "Response Get History Api History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CookSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nickname"
        ],
        "title": "CookSummary"
      },
...
      "DishCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "category": {
            "type": "string",
            "enum": [
              "荤菜",
              "素菜",
              "主食",
              "汤",
              "其他"
            ],
            "title": "Category"
          },
          "cook_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Cook Ids"
          },
          "ingredients": {
            "items": {
              "type": "string",
              "maxLength": 100,
              "minLength": 1
            },
            "type": "array",
            "minItems": 1,
            "title": "Ingredients"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Key"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "cook_ids",
          "ingredients"
        ],
        "title": "DishCreate"
      },
      "DishRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "enum": [
              "荤菜",
              "素菜",
              "主食",
              "汤",
              "其他"
            ],
            "title": "Category"
          },
          "cooks": {
            "items": {
              "$ref": "#/components/schemas/CookSummary"
            },
            "type": "array",
            "title": "Cooks"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientSummary"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Key"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "updated_by": {
            "$ref": "#/components/schemas/UpdatedBySummary"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "cooks",
          "ingredients",
          "image_key",
          "image_url",
          "archived_at",
          "updated_by",
          "updated_at"
        ],
        "title": "DishRead"
      },
      "DishUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "category": {
            "type": "string",
            "enum": [
              "荤菜",
              "素菜",
              "主食",
              "汤",
              "其他"
            ],
            "title": "Category"
          },
          "cook_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Cook Ids"
          },
          "ingredients": {
            "items": {
              "type": "string",
              "maxLength": 100,
              "minLength": 1
            },
            "type": "array",
            "minItems": 1,
            "title": "Ingredients"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Key"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "cook_ids",
          "ingredients"
        ],
        "title": "DishUpdate"
      },
      "EventCreate": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "meal_opened",
              "first_request_added",
              "menu_confirmed",
              "menu_modified"
            ],
            "title": "Name"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "EventCreate"
      },
      "EventRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "enum": [
              "meal_opened",
              "first_request_added",
              "menu_confirmed",
              "menu_modified"
            ],
            "title": "Name"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "properties",
          "created_at"
        ],
        "title": "EventRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryEntry": {
        "properties": {
          "meal_slot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meal Slot Id"
          },
          "local_date": {
            "type": "string",
            "format": "date",
            "title": "Local Date"
          },
          "meal_type": {
            "type": "string",
            "enum": [
              "lunch",
              "dinner"
            ],
            "title": "Meal Type"
          },
          "menu": {
            "items": {
              "$ref": "#/components/schemas/HistoryMenuItem"
            },
            "type": "array",
            "title": "Menu"
          },
          "last_modified_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistoryLastModifier"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Modified At"
          }
        },
        "type": "object",
        "required": [
          "meal_slot_id",
          "local_date",
          "meal_type",
          "menu",
          "last_modified_by",
          "last_modified_at"
        ],
        "title": "HistoryEntry"
      },
      "HistoryLastModifier": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nickname"
        ],
        "title": "HistoryLastModifier"
      },
      "HistoryMenuItem": {
        "properties": {
          "dish_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dish Id"
          },
          "dish_name": {
            "type": "string",
            "title": "Dish Name"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Key"
          }
        },
        "type": "object",
        "required": [
          "dish_id",
          "dish_name",
          "image_key"
        ],
        "title": "HistoryMenuItem"
      },
...
      "IngredientCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "aliases": {
            "items": {
              "type": "string",
              "maxLength": 100,
              "minLength": 1
            },
            "type": "array",
            "title": "Aliases"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "IngredientCreate"
      },
      "IngredientRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "aliases"
        ],
        "title": "IngredientRead"
      },
      "IngredientSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "IngredientSummary"
      },
...
      "UpdatedBySummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nickname"
        ],
        "title": "UpdatedBySummary"
      },
      "ValidationCheckinRead": {
        "properties": {
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "home_meal_count": {
            "type": "integer",
            "title": "Home Meal Count"
          },
          "offline_discussion_count": {
            "type": "integer",
            "title": "Offline Discussion Count"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "week_start",
          "home_meal_count",
          "offline_discussion_count",
          "updated_at"
        ],
        "title": "ValidationCheckinRead"
      },
      "ValidationCheckinWrite": {
        "properties": {
          "home_meal_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Home Meal Count"
          },
          "offline_discussion_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offline Discussion Count"
          }
        },
        "type": "object",
        "required": [
          "home_meal_count",
          "offline_discussion_count"
        ],
        "title": "ValidationCheckinWrite"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}